Split pipelines module in pipelines and pipeline_runs
